refactor(api): split key-rotation.service.ts into services/key-rotation/ (api-6fzx)
