+++ apps/api/src/__tests__/routes/buckets/rotations/claim.test.ts
@@ -12,10 +12,16 @@ import type { ApiErrorResponse } from "@pluralscape/types";
 
 // ── Mocks ────────────────────────────────────────────────────────
 
-vi.mock("../../../../services/key-rotation.service.js", () => ({
+vi.mock("../../../../services/key-rotation/initiate.js", () => ({
   initiateRotation: vi.fn(),
+}));
+vi.mock("../../../../services/key-rotation/claim.js", () => ({
   claimRotationChunk: vi.fn(),
+}));
+vi.mock("../../../../services/key-rotation/complete.js", () => ({
   completeRotationChunk: vi.fn(),
+}));
+vi.mock("../../../../services/key-rotation/queries.js", () => ({
   getRotationProgress: vi.fn(),
 }));
 
@@ -28,7 +34,7 @@ vi.mock("../../../../middleware/rate-limit.js", () => mockRateLimitFactory());
 vi.mock("../../../../middleware/auth.js", () => mockAuthFactory());
 // ── Imports after mocks ──────────────────────────────────────────
 
-const { claimRotationChunk } = await import("../../../../services/key-rotation.service.js");
+const { claimRotationChunk } = await import("../../../../services/key-rotation/claim.js");
 const { systemRoutes } = await import("../../../../routes/systems/index.js");
 
 // ── Helpers ──────────────────────────────────────────────────────
