+++ apps/api/src/routes/buckets/rotations/claim.ts
@@ -6,7 +6,7 @@ import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { claimRotationChunk } from "../../../services/key-rotation.service.js";
+import { claimRotationChunk } from "../../../services/key-rotation/claim.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
