+++ apps/api/src/routes/buckets/rotations/complete-chunk.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { completeRotationChunk } from "../../../services/key-rotation.service.js";
+import { completeRotationChunk } from "../../../services/key-rotation/complete.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
