+++ apps/api/src/routes/buckets/rotations/initiate.ts
@@ -9,7 +9,7 @@ import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createIdempotencyMiddleware } from "../../../middleware/idempotency.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { initiateRotation } from "../../../services/key-rotation.service.js";
+import { initiateRotation } from "../../../services/key-rotation/initiate.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
