+++ apps/api/src/routes/buckets/rotations/progress.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { getRotationProgress } from "../../../services/key-rotation.service.js";
+import { getRotationProgress } from "../../../services/key-rotation/queries.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
