+++ apps/api/src/routes/buckets/rotations/retry.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { retryRotation } from "../../../services/key-rotation.service.js";
+import { retryRotation } from "../../../services/key-rotation/retry.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
