+++ apps/api/src/trpc/routers/bucket.ts
@@ -34,13 +34,11 @@ import {
   restoreBucket,
   updateBucket,
 } from "../../services/bucket.service.js";
-import {
-  claimRotationChunk,
-  completeRotationChunk,
-  getRotationProgress,
-  initiateRotation,
-  retryRotation,
-} from "../../services/key-rotation.service.js";
+import { claimRotationChunk } from "../../services/key-rotation/claim.js";
+import { completeRotationChunk } from "../../services/key-rotation/complete.js";
+import { initiateRotation } from "../../services/key-rotation/initiate.js";
+import { getRotationProgress } from "../../services/key-rotation/queries.js";
+import { retryRotation } from "../../services/key-rotation/retry.js";
 import { createTRPCCategoryRateLimiter } from "../middlewares/rate-limit.js";
 import { systemProcedure } from "../middlewares/system.js";
 import { router } from "../trpc.js";
